refactor: move loan swap details into settings
