+++ apps/main/src/llamalend/widgets/action-card/LoanActionSettings.tsx
@@ -3,9 +3,18 @@ import type { MarketRoutes } from '@/llamalend/hooks/useMarketRoutes'
 import Collapse from '@mui/material/Collapse'
 import Stack from '@mui/material/Stack'
 import type { Decimal } from '@primitives/decimal.utils'
+import { maybe } from '@primitives/objects.utils'
 import { useSwitch } from '@ui-kit/hooks/useSwitch'
+import { t } from '@ui-kit/lib/i18n'
+import { ActionInfo } from '@ui-kit/shared/ui/ActionInfo'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
-import { borderStyle } from '@ui-kit/utils'
+import { mapQuery, type QueryProp } from '@ui-kit/types/util'
+import { borderStyle, decimal, formatNumber } from '@ui-kit/utils'
+import {
+  getPriceImpactDisplay,
+  getPriceImpactPercent,
+  type PriceImpact,
+} from '@ui-kit/widgets/DetailPageLayout/price-impact.util'
 import { RouteProvidersAccordion } from '@ui-kit/widgets/RouteProvider'
 import { SlippageToleranceActionInfo } from '@ui-kit/widgets/SlippageSettings'
 
@@ -16,13 +25,25 @@ export const LoanActionSettings = ({
   onSlippageChange,
   routes,
   show = true,
+  exchangeRate,
+  priceImpact,
+  collateralSymbol,
+  borrowSymbol,
 }: {
   slippage: Decimal | undefined
   onSlippageChange: (newSlippage: Decimal) => void
   routes?: MarketRoutes
   show?: boolean
+  exchangeRate?: QueryProp<Decimal | null>
+  priceImpact?: QueryProp<PriceImpact | Decimal | null>
+  collateralSymbol?: string
+  borrowSymbol?: string
 }) => {
   const [isRoutesOpen, , , toggleRoutes] = useSwitch(false)
+  const { label: priceImpactLabel, color: priceImpactColor } = getPriceImpactDisplay(priceImpact, {
+    slippage,
+    slippageType: LEVERAGE,
+  })
 
   return (
     <Collapse in={show}>
@@ -30,6 +51,7 @@ export const LoanActionSettings = ({
         data-testid="loan-action-settings"
         sx={{ backgroundColor: t => t.design.Layer[2].Fill, border: borderStyle, padding: Spacing.xs }}
       >
+        {routes && <RouteProvidersAccordion isExpanded={isRoutesOpen} onToggle={toggleRoutes} {...routes} />}
         {slippage && (
           <SlippageToleranceActionInfo
             maxSlippage={slippage}
@@ -38,7 +60,29 @@ export const LoanActionSettings = ({
             size="small"
           />
         )}
-        {routes && <RouteProvidersAccordion isExpanded={isRoutesOpen} onToggle={toggleRoutes} {...routes} />}
+        {exchangeRate && collateralSymbol && borrowSymbol && (
+          <ActionInfo
+            label={t`Exchange rate`}
+            value={mapQuery(exchangeRate, er =>
+              maybe(
+                decimal(er),
+                er =>
+                  `1 ${collateralSymbol} = ${formatNumber(er, { abbreviate: false, highPrecision: true })} ${borrowSymbol}`,
+              ),
+            )}
+            size="small"
+            testId="borrow-exchange-rate"
+          />
+        )}
+        {priceImpact && (
+          <ActionInfo
+            label={priceImpactLabel}
+            value={mapQuery(priceImpact, data => formatNumber(getPriceImpactPercent(data), 'percent.rate'))}
+            valueColor={priceImpactColor}
+            size="small"
+            testId="borrow-price-impact"
+          />
+        )}
       </Stack>
     </Collapse>
   )
